fix: change path to assets, rename files and folders, add .idea in gitignore
