+++ src/core/components/sprint-game/sprint-game-view/sprint-game.css
