+++ webpack.config.js
@@ -4,7 +4,7 @@ const { merge } = require('webpack-merge');
 const HtmlWebpackPlugin = require('html-webpack-plugin');
 const { CleanWebpackPlugin } = require('clean-webpack-plugin');
 const EslintPlugin = require('eslint-webpack-plugin');
-const CopyPlugin = require("copy-webpack-plugin");
+const CopyPlugin = require('copy-webpack-plugin');
 
 const baseConfig = {
     entry: path.resolve(__dirname, './src/index'),
@@ -38,6 +38,7 @@ const baseConfig = {
     output: {
         filename: 'index.js',
         path: path.resolve(__dirname, './dist'),
+        assetModuleFilename: 'assets/[name][ext]',
     },
     plugins: [
         new webpack.ProvidePlugin({
@@ -50,7 +51,7 @@ const baseConfig = {
         }),
         new CleanWebpackPlugin(),
         new CopyPlugin({
-            patterns: [{ from: './src/core/components/sprintGame/assets/svg', to: '../dist/assets/svg' }],
+            patterns: [{ from: './src/assets/svg', to: '../dist/assets/svg' }],
         }),
     ],
 };
